apply changes from code-review

+++ learning_loop_node/trainer/trainer_logic.py
@@ -55,6 +55,7 @@ class TrainerLogic():
         self._active_training_io: Optional[ActiveTrainingIO] = None
         self._node: Optional[TrainerNode] = None
         self.restart_after_training = bool(int(os.environ.get('RESTART_AFTER_TRAINING', '0')))
+        self.keep_old_trainings = bool(int(os.environ.get('KEEP_OLD_TRAININGS', '0')))
         self.inference_batch_size = int(os.environ.get('INFERENCE_BATCH_SIZE', '10'))
         logging.info(f'INFERENCE_BATCH_SIZE: {self.inference_batch_size}')
 
@@ -89,8 +90,9 @@ class TrainerLogic():
         self._node = node
         try:
             project_folder = Node.create_project_folder(context)
-            # NOTE: We delete all existing training folders because they are not needed anymore.
-            TrainerLogic.delete_all_training_folders(project_folder)
+            if not self.keep_old_trainings:
+                # NOTE: We delete all existing training folders because they are not needed anymore.
+                TrainerLogic.delete_all_training_folders(project_folder)
             self._training = TrainerLogic.generate_training(project_folder, context)
             self._training.data = TrainingData(categories=Category.from_list(details['categories']))
             self._training.data.hyperparameter = from_dict(data_class=Hyperparameter, data=details)
@@ -159,7 +161,7 @@ class TrainerLogic():
                 await self.upload_detections()
             elif tstate == TrainingState.ReadyForCleanup:
                 await self.clear_training()
-                self.restart()
+                self.may_restart()
 
     def load_last_training(self) -> None:
         self._training = self.node.last_training_io.load()
@@ -522,7 +524,7 @@ class TrainerLogic():
                 except asyncio.CancelledError:
                     pass
                 logging.info('cancelled training task')
-                self.restart()
+                self.may_restart()
 
     async def shutdown(self) -> None:
         self.shutdown_event.set()
@@ -532,7 +534,7 @@ class TrainerLogic():
     def get_log(self) -> str:
         return self.executor.get_log()
 
-    def restart(self) -> None:
+    def may_restart(self) -> None:
         if self.restart_after_training:
             logging.info('restarting')
             assert self._node is not None
